move support file out of lib
because it should only be used for extending test-unit
