+++ test/error_lifecycle.rb
